Updates and Resolved Issue Issue - 1 - In CartActivity when products were present in the local DB, for multiple Stores, then due to only wrapping the Storing of Products fetched from the Local DB for Current Store_ID in if condition to add only those products to CartActivity whose Store_ID is Currently selected, made the Total_Amt. for the products present of other stores in local DB, get added in the CartActivity for the Current Store_ID. Fix - extended the If condition for all the operation done while retrieving the Products from the Local DB.
1 - DBHelper
1_1 - added a new Column Has_Image, in the Products_Table.

Issue - 2 - after adding the checks for ensuring that the Rs.100 Bonus banner is shown for the First time the Users Registers to the App, but when a User Uninstalls the app and re-installs it, the Banner was showed to the User again, as the App was opened for the First time acc. to the App.
Fix - added Checks i.e., isDirectLogin to ensure that the Banner is Shown to the user only when, the the App is Opened First time and There is no DirectLogin(i.e., the User already existed in the DB) to the App.
2 - MyProfile(i.e., Home)
2_1 - added isDirectLogin Check to resolve the Above Issue.
3 - Covid19 Activity
3_1 - added check for where the Intent to this Activity is coming from, If MyProfile(i.e., Home) then set the isDirectLogin flag to true, otherwise set it to false.
4 - Covid19, OTPConfirmActivity, MainActivity.
4_1 - added activity flag to Intent to Identify DirectLogin.

Issue - 3 -  if there are same Products stored in the Local DB for different Store_ID's (such as for 1 and 2) then, User's will only be able to see the Product of that Store_ID, which they had added to the local DB first, and after that, if you try to add that same product with another Store_ID, the quantity was getting Updated in the product of the first Store_ID that was stored.
Fix - Before the Product is Inserted in the Database, its existence is checked in the Database(was previously checked by the barcode only, which caused this Issue). So, now added Store_ID with the barcode to check if the Product already exists in the Database for that Store_ID too.
5 - DBHelper
5_1 - added extra parameters Store_ID in product_exists() and update_product() Funcs.
6 - ProductDetails
6_1 - added store_id as parameters to Add_To_Basket() Func.

7 - Product.java
7_1 - added new parameter hasImage to the Class.
8 - ProductAdapter
8_1 - added check, if the product has Image then only show the image.
9 - StoresAdapter
9_1 - removed constrains for store_ID to be Accessible to the User, now all Stores are available.
10 - TxnDetails
10_1 - added has_image parameter to the Product Class.